src/requirements.txt
asgiref==3.2.7
astroid==2.4.1
beautifulsoup4==4.9.1
colorama==0.4.3
Django==3.0.6
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.1.1
django-multiselectfield==0.1.12
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.5
pylint==2.5.2
python-dotenv==0.13.0
pytz==2020.1
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
wrapt==1.12.1
